[IMP] account_payment_kmitl: refactor payment batch to payment cycle
Replace payment.batch (workflow model with tier validation) with
payment.cycle (simple master data). Add payment_cycle_id to both
account.payment and bank.payment.export for categorizing payments
by payment round.

Remove tier validation, security groups, and PDF report that are
no longer needed.
