+++ account_payment_kmitl/__manifest__.py
@@ -12,7 +12,6 @@
         "account",
         "account_move_kmitl",
         "account_fiscal_year_enhance",
-        "base_tier_validation",
         "budget",
         "l10n_th_bank_payment_export",
         "l10n_th_bank_payment_export_format",
@@ -22,12 +21,10 @@
         "l10n_th_account_wht_cert_form"
     ],
     "data": [
-        "security/payment_batch_security.xml",
         "security/ir.model.access.csv",
         "data/kmitl_payment_type_data.xml",
         "data/tier_definition.xml",
         "data/payment_batch_sequence.xml",
-        "data/payment_batch_tier_definition.xml",
         "views/kmitl_payment_type_views.xml",
         "views/account_payment_views.xml",
         "views/bank_payment_export_views.xml",
@@ -36,8 +33,6 @@
         "report/paperformat.xml",
         "report/report_bank_payment_export_action.xml",
         "report/report_bank_payment_export.xml",
-        "report/report_payment_batch_action.xml",
-        "report/report_payment_batch.xml",
     ],
     "installable": True,
     "application": True,
