+++ account_payment_kmitl/data/payment_batch_sequence.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo noupdate="1">
 
-    <record id="seq_payment_batch" model="ir.sequence">
-        <field name="name">Payment Batch</field>
-        <field name="code">payment.batch</field>
-        <field name="prefix">PB/%(range_year)s/</field>
+    <record id="seq_payment_cycle" model="ir.sequence">
+        <field name="name">Payment Cycle</field>
+        <field name="code">payment.cycle</field>
+        <field name="prefix">CYC/%(range_year)s/</field>
         <field name="padding">3</field>
         <field name="number_next">1</field>
         <field name="number_increment">1</field>
