--- account_payment_kmitl/data/payment_batch_tier_definition.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo noupdate="1">
-
-    <record id="tier_definition_payment_batch" model="tier.definition">
-        <field name="model_id" ref="model_payment_batch"/>
-        <field name="name">อนุมัติรอบการจ่ายเงิน</field>
-        <field name="sequence">10</field>
-        <field name="review_type">group</field>
-        <field name="reviewer_group_id" ref="group_payment_batch_manager"/>
-        <field name="definition_type">domain</field>
-        <field name="definition_domain">[]</field>
-        <field name="approve_sequence">True</field>
-        <field name="has_comment">False</field>
-        <field name="active">True</field>
-        <field name="notify_on_create">True</field>
-        <field name="notify_on_accepted">True</field>
-        <field name="notify_on_rejected">True</field>
-        <field name="notify_on_restarted">True</field>
-    </record>
-
-</odoo>
