--- account_payment_kmitl/models/tier_definition.py
@@ -1,13 +1,0 @@
-# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
-
-from odoo import api, models
-
-
-class TierDefinition(models.Model):
-    _inherit = "tier.definition"
-
-    @api.model
-    def _get_tier_validation_model_names(self):
-        res = super()._get_tier_validation_model_names()
-        res.append("payment.batch")
-        return res
