--- account_payment_kmitl/security/payment_batch_security.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="module_category_payment_batch" model="ir.module.category">
-        <field name="name">Payment Batch</field>
-        <field name="parent_id" ref="base.module_category_accounting_accounting"/>
-        <field name="sequence">25</field>
-    </record>
-
-    <!-- Finance Officer: create/edit/submit payment batch -->
-    <record id="group_payment_batch_officer" model="res.groups">
-        <field name="name">Finance Officer</field>
-        <field name="category_id" ref="module_category_payment_batch"/>
-        <field name="comment">เจ้าหน้าที่การเงิน สร้าง แก้ไข และส่งรอบการจ่ายเงินเพื่ออนุมัติ</field>
-    </record>
-
-    <!-- Finance Manager: approve payment batch (implies officer) -->
-    <record id="group_payment_batch_manager" model="res.groups">
-        <field name="name">Finance Manager</field>
-        <field name="category_id" ref="module_category_payment_batch"/>
-        <field name="implied_ids" eval="[(4, ref('group_payment_batch_officer'))]"/>
-        <field name="comment">ผู้มีอำนาจอนุมัติรอบการจ่ายเงิน</field>
-    </record>
-
-</odoo>
